+++ app/actions/install/install_scripts.go
